+++ resources/ui5/library-dbg.js
@@ -15,7 +15,7 @@ sap.ui.define([
 	// Hint: sap.ui.getCore() must still be used to support preload with sync bootstrap!
 	sap.ui.getCore().initLibrary({
 		name: "com.laravelui5.core",
-		version: "2.3.4",
+		version: "2.3.5",
 		dependencies: [
 			// keep in sync with the ui5.yaml and .library files
 			"sap.ui.core"
@@ -27,6 +27,7 @@ sap.ui.define([
 			"com.laravelui5.core.BaseComponent",
 			"com.laravelui5.core.BaseController",
 			"com.laravelui5.core.Connection",
+			"com.laravelui5.core.Context",
 			"com.laravelui5.core.LaravelUi5"
 		],
 		noLibraryCSS: true // if no CSS is provided, you can disable the library.css load here
@@ -39,7 +40,7 @@ sap.ui.define([
 	 * @namespace
 	 * @alias com.laravelui5.core
 	 * @author Michael Gerzabek
-	 * @version 2.3.4
+	 * @version 2.3.5
 	 * @public
 	 */
 	return ObjectPath.get("com.laravelui5.core");