+++ src/DashboardModule.php
@@ -36,7 +36,7 @@ class DashboardModule extends AbstractUi5Module
         return null;
     }
 
-    public function getArtifactRoot(): Ui5ArtifactInterface
+    public function getArtifactRoot(): Ui5AppInterface|Ui5LibraryInterface
     {
         return $this->getApp();
     }