+++ src/ReportModule.php
@@ -35,7 +35,7 @@ class ReportModule extends AbstractUi5Module
         return null;
     }
 
-    public function getArtifactRoot(): Ui5ArtifactInterface
+    public function getArtifactRoot(): Ui5AppInterface|Ui5LibraryInterface
     {
         return $this->getApp();
     }