+++ tests-fixture/src/HelloModule.php
@@ -29,7 +29,7 @@ class HelloModule extends AbstractUi5Module
         return null;
     }
 
-    public function getArtifactRoot(): Ui5ArtifactInterface
+    public function getArtifactRoot(): Ui5AppInterface|Ui5LibraryInterface
     {
         return $this->getApp();
     }